+++ plutarch-test/goldens/rec.bench.golden
@@ -1,0 +1,45 @@
+simple.constr.pcon {"exBudgetCPU":29873,"exBudgetMemory":200,"scriptSizeBytes":30}
+simple.constr.pcon' {"exBudgetCPU":29873,"exBudgetMemory":200,"scriptSizeBytes":30}
+simple.constr.rcon {"exBudgetCPU":29873,"exBudgetMemory":200,"scriptSizeBytes":30}
+simple.field.access-term {"exBudgetCPU":387149,"exBudgetMemory":1400,"scriptSizeBytes":33}
+simple.field.value.direct-access {"exBudgetCPU":387149,"exBudgetMemory":1400,"scriptSizeBytes":33}
+simple.field.value.pmatch {"exBudgetCPU":387149,"exBudgetMemory":1400,"scriptSizeBytes":33}
+simple.field.value.pmatch' {"exBudgetCPU":387149,"exBudgetMemory":1400,"scriptSizeBytes":33}
+simple.field.value.rmatch {"exBudgetCPU":387149,"exBudgetMemory":1400,"scriptSizeBytes":33}
+simple.reconstr.pcon {"exBudgetCPU":387149,"exBudgetMemory":1400,"scriptSizeBytes":39}
+simple.reconstr.field-access {"exBudgetCPU":744425,"exBudgetMemory":2600,"scriptSizeBytes":43}
+LetRec.record {"exBudgetCPU":655106,"exBudgetMemory":2300,"scriptSizeBytes":52}
+LetRec.record-field {"exBudgetCPU":655106,"exBudgetMemory":2300,"scriptSizeBytes":52}
+LetRec.even {"exBudgetCPU":774198,"exBudgetMemory":2700,"scriptSizeBytes":82}
+LetRec.even.4 {"exBudgetCPU":7105378,"exBudgetMemory":19018,"scriptSizeBytes":85}
+LetRec.even.5 {"exBudgetCPU":8542057,"exBudgetMemory":22722,"scriptSizeBytes":85}
+nested.flat.reconstr-with-rcon {"exBudgetCPU":119192,"exBudgetMemory":500,"scriptSizeBytes":60}
+nested.flat.nested-field-access {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":67}
+nested.flat.nested-field-value.direct-access {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":67}
+nested.flat.nested-field-value.pmatch {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":67}
+nested.flat.nested-field-value.pmatch' {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":67}
+nested.flat.nested-field-value.rmatch {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":67}
+nested.flat.reconstr-with-pcon {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":85}
+nested.flat.reconstr-nested-field-value {"exBudgetCPU":1905572,"exBudgetMemory":6500,"scriptSizeBytes":91}
+nested.flat.nested-record-access-term {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":73}
+nested.flat.nested-match-term {"exBudgetCPU":1369658,"exBudgetMemory":4700,"scriptSizeBytes":77}
+nested.flat.nested-match-value {"exBudgetCPU":1369658,"exBudgetMemory":4700,"scriptSizeBytes":77}
+nested.shallow.constr-with-rcon {"exBudgetCPU":119192,"exBudgetMemory":500,"scriptSizeBytes":65}
+nested.shallow.nested-field-access {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":73}
+nested.shallow.nested-field-value.direct-access {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":73}
+nested.shallow.nested-field-value.pmatch {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":73}
+nested.shallow.nested-field-value.pmatch' {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":73}
+nested.shallow.nested-field-value.rmatch {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":73}
+nested.shallow.reconstr-with-pcon {"exBudgetCPU":655106,"exBudgetMemory":2300,"scriptSizeBytes":80}
+nested.shallow.reconstr-nested-field-value {"exBudgetCPU":1548296,"exBudgetMemory":5300,"scriptSizeBytes":88}
+nested.shallow.nested-record-access-term {"exBudgetCPU":655106,"exBudgetMemory":2300,"scriptSizeBytes":70}
+nested.shallow.nested-match-term {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":73}
+nested.shallow.nested-match-value {"exBudgetCPU":1012382,"exBudgetMemory":3500,"scriptSizeBytes":73}
+Data.pdata.simple {"exBudgetCPU":2671978,"exBudgetMemory":5329,"scriptSizeBytes":70}
+Data.pdata.simple-value-deconstructed {"exBudgetCPU":2881524,"exBudgetMemory":5561,"scriptSizeBytes":72}
+Data.pdata.flat-data-deconstructed {"exBudgetCPU":7537231,"exBudgetMemory":14647,"scriptSizeBytes":169}
+Data.pdata.shallow-data-deconstructed {"exBudgetCPU":9030421,"exBudgetMemory":17775,"scriptSizeBytes":200}
+Data.fieldFromData.simple {"exBudgetCPU":29873,"exBudgetMemory":200,"scriptSizeBytes":36}
+Data.fieldFromData.nested.flat {"exBudgetCPU":148965,"exBudgetMemory":600,"scriptSizeBytes":48}
+Data.fieldFromData.nested.shallow {"exBudgetCPU":684879,"exBudgetMemory":2400,"scriptSizeBytes":115}
+Data.pfromData {"exBudgetCPU":565787,"exBudgetMemory":2000,"scriptSizeBytes":83}
