+++ plutarch-test/goldens/rec.uplc.eval.golden
@@ -1,0 +1,59 @@
+simple.constr.pcon (program 1.0.0 (\i0 -> i1 False 6 "Salut, Monde!"))
+simple.constr.pcon' (program 1.0.0 (\i0 -> i1 False 6 "Salut, Monde!"))
+simple.constr.rcon (program 1.0.0 (\i0 -> i1 False 6 "Salut, Monde!"))
+simple.field.access-term (program 1.0.0 6)
+simple.field.value.direct-access (program 1.0.0 6)
+simple.field.value.pmatch (program 1.0.0 "Salut, Monde!")
+simple.field.value.pmatch' (program 1.0.0 "Salut, Monde!")
+simple.field.value.rmatch (program 1.0.0 "Salut, Monde!")
+simple.reconstr.pcon (program 1.0.0 (\i0 -> i1 False 6 "Salut, Monde!"))
+simple.reconstr.field-access (program 1.0.0 6)
+LetRec.record (program 1.0.0 12)
+LetRec.record-field (program 1.0.0 12)
+LetRec.even (program 1.0.0 (\i0 -> force (force ifThenElse (equalsInteger i1 0) (delay True) (delay ((\i0 -> (\i0 -> (\i0 -> \i0 -> i1 (\i0 -> force (force ifThenElse (equalsInteger i1 0) (delay True) (delay (i3 (\i0 -> \i0 -> i1) (subtractInteger i1 1))))) (\i0 -> force (force ifThenElse (equalsInteger i1 0) (delay False) (delay (i3 (\i0 -> \i0 -> i2) (subtractInteger i1 1)))))) (\i0 -> i2 i2 i1)) (\i0 -> (\i0 -> \i0 -> i1 (\i0 -> force (force ifThenElse (equalsInteger i1 0) (delay True) (delay (i3 (\i0 -> \i0 -> i1) (subtractInteger i1 1))))) (\i0 -> force (force ifThenElse (equalsInteger i1 0) (delay False) (delay (i3 (\i0 -> \i0 -> i2) (subtractInteger i1 1)))))) (\i0 -> i2 i2 i1)) i1) (\i0 -> \i0 -> i1) (subtractInteger i1 1))))))
+LetRec.even.4 (program 1.0.0 True)
+LetRec.even.5 (program 1.0.0 False)
+nested.flat.reconstr-with-rcon (program 1.0.0 (\i0 -> i1 False False 6 "Salut, Monde!" 4 False 9 "Salut, Monde!" "Hola, Mundo!"))
+nested.flat.nested-field-access (program 1.0.0 9)
+nested.flat.nested-field-value.direct-access (program 1.0.0 9)
+nested.flat.nested-field-value.pmatch (program 1.0.0 9)
+nested.flat.nested-field-value.pmatch' (program 1.0.0 "Salut, Monde!")
+nested.flat.nested-field-value.rmatch (program 1.0.0 "Salut, Monde!")
+nested.flat.reconstr-with-pcon (program 1.0.0 (\i0 -> i1 False False 6 "Salut, Monde!" 4 False 9 "Salut, Monde!" "Hola, Mundo!"))
+nested.flat.reconstr-nested-field-value (program 1.0.0 9)
+nested.flat.nested-record-access-term (program 1.0.0 (\i0 -> i1 False 6 "Salut, Monde!"))
+nested.flat.nested-match-term (program 1.0.0 "Salut, Monde!")
+nested.flat.nested-match-value (program 1.0.0 "Salut, Monde!")
+nested.shallow.constr-with-rcon (program 1.0.0 (\i0 -> i1 False (\i0 -> i1 False 6 "Salut, Monde!") 4 (\i0 -> i1 False 9 "Salut, Monde!") "Hola, Mundo!"))
+nested.shallow.nested-field-access (program 1.0.0 9)
+nested.shallow.nested-field-value.direct-access (program 1.0.0 9)
+nested.shallow.nested-field-value.pmatch (program 1.0.0 9)
+nested.shallow.nested-field-value.pmatch' (program 1.0.0 "Salut, Monde!")
+nested.shallow.nested-field-value.rmatch (program 1.0.0 "Salut, Monde!")
+nested.shallow.reconstr-with-pcon (program 1.0.0 (\i0 -> i1 False (\i0 -> i1 False 6 "Salut, Monde!") 4 (\i0 -> i1 False 9 "Salut, Monde!") "Hola, Mundo!"))
+nested.shallow.reconstr-nested-field-value (program 1.0.0 9)
+nested.shallow.nested-record-access-term (program 1.0.0 (\i0 -> i1 False 6 "Salut, Monde!"))
+nested.shallow.nested-match-term (program 1.0.0 "Salut, Monde!")
+nested.shallow.nested-match-value (program 1.0.0 "Salut, Monde!")
+Data.pdata.simple (program 1.0.0 #d8799fd87980064d53616c75742c204d6f6e646521ff)
+Data.pdata.simple-value-deconstructed (program 1.0.0 (0, [#d87980, #06, #4d53616c75742c204d6f6e646521]))
+Data.pdata.flat-data-deconstructed (program 1.0.0 ( 0
+, [ #d87980
+  , #d87980
+  , #06
+  , #4d53616c75742c204d6f6e646521
+  , #04
+  , #d87980
+  , #09
+  , #4d53616c75742c204d6f6e646521
+  , #4c486f6c612c204d756e646f21 ] ))
+Data.pdata.shallow-data-deconstructed (program 1.0.0 ( 0
+, [ #d87980
+  , #d8799fd87980064d53616c75742c204d6f6e646521ff
+  , #04
+  , #d8799fd87980094d53616c75742c204d6f6e646521ff
+  , #4c486f6c612c204d756e646f21 ] ))
+Data.fieldFromData.simple (program 1.0.0 (\i0 -> unIData ((\i0 -> force (force ifThenElse (equalsInteger (force (force fstPair) i1) 0) (delay (force headList (force tailList (force (force sndPair) i1)))) (delay error))) (unConstrData i1))))
+Data.fieldFromData.nested.flat (program 1.0.0 (\i0 -> unIData ((\i0 -> force (force ifThenElse (equalsInteger (force (force fstPair) i1) 0) (delay (force headList (force tailList (force tailList (force tailList (force tailList (force tailList (force tailList (force (force sndPair) i1))))))))) (delay error))) (unConstrData i1))))
+Data.fieldFromData.nested.shallow (program 1.0.0 (\i0 -> (\i0 -> (\i0 -> force (force ifThenElse (equalsInteger (force (force fstPair) i1) 0) (delay (\i0 -> i1 ((\i0 -> equalsInteger (force (force fstPair) (unConstrData i1)) 1) (force headList (force (force sndPair) i2))) (unIData (force headList (force tailList (force (force sndPair) i2)))) (decodeUtf8 (unBData (force headList (force tailList (force tailList (force (force sndPair) i2)))))))) (delay error))) (unConstrData i1)) ((\i0 -> force (force ifThenElse (equalsInteger (force (force fstPair) i1) 0) (delay (force headList (force tailList (force tailList (force tailList (force (force sndPair) i1)))))) (delay error))) (unConstrData i1)) (\i0 -> \i0 -> \i0 -> i2)))
+Data.pfromData (program 1.0.0 (\i0 -> (\i0 -> force (force ifThenElse (equalsInteger (force (force fstPair) i1) 0) (delay (\i0 -> i1 ((\i0 -> equalsInteger (force (force fstPair) (unConstrData i1)) 1) (force headList (force (force sndPair) i2))) (unIData (force headList (force tailList (force (force sndPair) i2)))) (decodeUtf8 (unBData (force headList (force tailList (force tailList (force (force sndPair) i2)))))))) (delay error))) (unConstrData i1) (\i0 -> \i0 -> \i0 -> i2)))
