+++ plutarch-test/plutarch-test.cabal
@@ -108,6 +108,7 @@ executable plutarch-test
   build-depends:
     , base16-bytestring
     , cborg
+    , rank2classes
     , serialise
 
   if impl(ghc >=9.0)
@@ -126,6 +127,7 @@ executable plutarch-test
     Plutarch.PLamSpec
     Plutarch.PlutusTypeSpec
     Plutarch.RationalSpec
+    Plutarch.RecSpec
     Plutarch.RecursionSpec
     Plutarch.ScriptsSpec
     Plutarch.StringSpec
